pyproject.toml
[project]
name = "langchainapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
    "anthropic>=0.79.0",
    "chainlit>=2.9.6",
    "chromadb>=1.5.0",
    "fastapi>=0.128.6",
    "langchain>=1.2.9",
    "langchain-anthropic>=1.3.2",
    "langchain-core>=1.2.9",
    "langchain-openai>=1.1.8",
    "langgraph>=1.0.8",
    "langsmith>=0.7.1",
    "literalai>=0.1.201",
    "openai>=2.18.0",
    "openai-agents>=0.8.3",
    "opentelemetry-api>=1.39.1",
    "pandas>=3.0.0",
    "pgai>=0.12.1",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
